Update Unit tests
