+++ Frontend.Tests/PagesTests/TaskList/HtbDocument/KeyStage2PerformanceTests.cs
@@ -3,53 +3,35 @@ using Data;
 using Data.Models;
 using Data.Models.KeyStagePerformance;
 using Frontend.Models;
-using Frontend.Pages.TaskList.KeyStage4Performance;
-using Frontend.Services.Interfaces;
-using Frontend.Services.Responses;
+using Frontend.Pages.TaskList.KeyStage2Performance;
 using Frontend.Tests.Helpers;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Moq;
 using Xunit;
 
-namespace Frontend.Tests.PagesTests.TaskList
+namespace Frontend.Tests.PagesTests.TaskList.HtbDocument
 {
-    public class KeyStage4PerformanceTests : PageTests
+    public class KeyStage2PerformanceTests : PageTests
     {
-        private readonly KeyStage4Performance _subject;
+        private readonly KeyStage2Performance _subject;
 
-        public KeyStage4PerformanceTests()
+        public KeyStage2PerformanceTests()
         {
-            //arrange
             FoundInformationForProject.EducationPerformance = new EducationPerformance
             {
-                KeyStage4Performance = new List<KeyStage4>
+                KeyStage2Performance = new List<KeyStage2>
                 {
-                    new KeyStage4
+                    new KeyStage2
                     {
-                        Year = "2019-2020",
-                        SipNumberofpupilsprogress8 = new DisadvantagedPupilsResult
-                        {
-                            NotDisadvantaged = "20.5",
-                            Disadvantaged = "10.5"
-                        }
-                    },
-                    new KeyStage4
-                    {
-                        Year = "2018-2019",
-                        SipNumberofpupilsprogress8 = new DisadvantagedPupilsResult
-                        {
-                            NotDisadvantaged = "40.8",
-                            Disadvantaged = "30.4"
-                        }
+                        Year = "test year"
                     }
                 }
             };
-
-            _subject = new KeyStage4Performance(GetInformationForProject.Object, ProjectRepository.Object);
+            _subject = new KeyStage2Performance(GetInformationForProject.Object, ProjectRepository.Object);
         }
 
-        public class OnGetAsyncTests : KeyStage4PerformanceTests
+        public class OnGetAsyncTests : KeyStage2PerformanceTests
         {
             [Fact]
             public async void GivenUrn_FetchesProjectFromTheRepository()
@@ -67,16 +49,14 @@ namespace Frontend.Tests.PagesTests.TaskList
                 Assert.IsType<PageResult>(response);
                 Assert.Equal(ProjectUrn0001, _subject.ProjectUrn);
                 Assert.Equal(AcademyUrn, _subject.OutgoingAcademyUrn);
-                Assert.Equal(2, _subject.EducationPerformance.KeyStage4Performance.Count);
-                Assert.Equal("2019-2020", _subject.EducationPerformance.KeyStage4Performance[0].Year);
-                Assert.Equal("2018-2019", _subject.EducationPerformance.KeyStage4Performance[1].Year);
+                Assert.Equal("test year", _subject.EducationPerformance.KeyStage2Performance[0].Year);
             }
 
             [Fact]
             public async void GivenAdditionalInformation_UpdatesTheViewModel()
             {
                 const string additionalInformation = "some additional info";
-                FoundInformationForProject.Project.KeyStage4PerformanceAdditionalInformation = additionalInformation;
+                FoundInformationForProject.Project.KeyStage2PerformanceAdditionalInformation = additionalInformation;
                 GetInformationForProject.Setup(s => s.Execute(ProjectUrn0001))
                     .ReturnsAsync(FoundInformationForProject);
 
@@ -85,7 +65,7 @@ namespace Frontend.Tests.PagesTests.TaskList
                 Assert.Equal(additionalInformation, _subject.AdditionalInformation.AdditionalInformation);
                 Assert.Equal(ProjectUrn0001, _subject.AdditionalInformation.Urn);
             }
-
+            
             [Fact]
             public async void GivenReturnToPreview_UpdatesTheViewModel()
             {
@@ -98,7 +78,7 @@ namespace Frontend.Tests.PagesTests.TaskList
             public async void GivenGetByUrnReturnsError_DisplayErrorPage()
             {
                 var pageModel =
-                    RazorPageTestHelpers.GetPageModelWithViewData<KeyStage4Performance>(
+                    RazorPageTestHelpers.GetPageModelWithViewData<KeyStage2Performance>(
                         GetInformationForProject.Object, ProjectRepository.Object);
 
                 var response = await pageModel.OnGetAsync(ProjectErrorUrn);
@@ -109,7 +89,7 @@ namespace Frontend.Tests.PagesTests.TaskList
             }
         }
 
-        public class OnPostAsyncTests : KeyStage4PerformanceTests
+        public class OnPostAsyncTests : KeyStage2PerformanceTests
         {
             private readonly Project _foundProject;
 
@@ -148,7 +128,7 @@ namespace Frontend.Tests.PagesTests.TaskList
             public async void GivenGetByUrnReturnsError_DisplayErrorPage()
             {
                 var pageModel =
-                    RazorPageTestHelpers.GetPageModelWithViewData<KeyStage4Performance>(
+                    RazorPageTestHelpers.GetPageModelWithViewData<KeyStage2Performance>(
                         GetInformationForProject.Object, ProjectRepository.Object);
 
                 var response = await pageModel.OnPostAsync(ProjectErrorUrn, string.Empty, false);
@@ -166,9 +146,9 @@ namespace Frontend.Tests.PagesTests.TaskList
                 var response = await _subject.OnPostAsync(ProjectUrn0001, additionalInformation, false);
 
                 var redirectToPageResponse = Assert.IsType<RedirectToPageResult>(response);
-                Assert.Equal("KeyStage4Performance", redirectToPageResponse.PageName);
+                Assert.Equal("KeyStage2Performance", redirectToPageResponse.PageName);
                 Assert.Equal("OnGetAsync", redirectToPageResponse.PageHandler);
-                Assert.Equal(additionalInformation, _foundProject.KeyStage4PerformanceAdditionalInformation);
+                Assert.Equal(additionalInformation, _foundProject.KeyStage2PerformanceAdditionalInformation);
             }
 
             [Fact]
@@ -178,10 +158,10 @@ namespace Frontend.Tests.PagesTests.TaskList
 
                 await _subject.OnPostAsync(ProjectUrn0001, additionalInfo, false);
                 ProjectRepository.Verify(r => r.Update(It.Is<Project>(
-                    project => project.KeyStage4PerformanceAdditionalInformation == additionalInfo
+                    project => project.KeyStage2PerformanceAdditionalInformation == additionalInfo
                 )));
             }
-
+            
             [Fact]
             public async void GivenReturnToPreview_RedirectsToThePreviewPage()
             {
