+++ Frontend.Tests/PagesTests/TaskList/HtbDocument/KeyStage4PerformanceTests.cs
@@ -3,37 +3,51 @@ using Data;
 using Data.Models;
 using Data.Models.KeyStagePerformance;
 using Frontend.Models;
-using Frontend.Pages.TaskList.KeyStage2Performance;
-using Frontend.Services.Interfaces;
-using Frontend.Services.Responses;
+using Frontend.Pages.TaskList.KeyStage4Performance;
 using Frontend.Tests.Helpers;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Moq;
 using Xunit;
 
-namespace Frontend.Tests.PagesTests.TaskList
+namespace Frontend.Tests.PagesTests.TaskList.HtbDocument
 {
-    public class KeyStage2PerformanceTests : PageTests
+    public class KeyStage4PerformanceTests : PageTests
     {
-        private readonly KeyStage2Performance _subject;
+        private readonly KeyStage4Performance _subject;
 
-        public KeyStage2PerformanceTests()
+        public KeyStage4PerformanceTests()
         {
+            //arrange
             FoundInformationForProject.EducationPerformance = new EducationPerformance
             {
-                KeyStage2Performance = new List<KeyStage2>
+                KeyStage4Performance = new List<KeyStage4>
                 {
-                    new KeyStage2
+                    new KeyStage4
                     {
-                        Year = "test year"
+                        Year = "2019-2020",
+                        SipNumberofpupilsprogress8 = new DisadvantagedPupilsResult
+                        {
+                            NotDisadvantaged = "20.5",
+                            Disadvantaged = "10.5"
+                        }
+                    },
+                    new KeyStage4
+                    {
+                        Year = "2018-2019",
+                        SipNumberofpupilsprogress8 = new DisadvantagedPupilsResult
+                        {
+                            NotDisadvantaged = "40.8",
+                            Disadvantaged = "30.4"
+                        }
                     }
                 }
             };
-            _subject = new KeyStage2Performance(GetInformationForProject.Object, ProjectRepository.Object);
+
+            _subject = new KeyStage4Performance(GetInformationForProject.Object, ProjectRepository.Object);
         }
 
-        public class OnGetAsyncTests : KeyStage2PerformanceTests
+        public class OnGetAsyncTests : KeyStage4PerformanceTests
         {
             [Fact]
             public async void GivenUrn_FetchesProjectFromTheRepository()
@@ -51,14 +65,16 @@ namespace Frontend.Tests.PagesTests.TaskList
                 Assert.IsType<PageResult>(response);
                 Assert.Equal(ProjectUrn0001, _subject.ProjectUrn);
                 Assert.Equal(AcademyUrn, _subject.OutgoingAcademyUrn);
-                Assert.Equal("test year", _subject.EducationPerformance.KeyStage2Performance[0].Year);
+                Assert.Equal(2, _subject.EducationPerformance.KeyStage4Performance.Count);
+                Assert.Equal("2019-2020", _subject.EducationPerformance.KeyStage4Performance[0].Year);
+                Assert.Equal("2018-2019", _subject.EducationPerformance.KeyStage4Performance[1].Year);
             }
 
             [Fact]
             public async void GivenAdditionalInformation_UpdatesTheViewModel()
             {
                 const string additionalInformation = "some additional info";
-                FoundInformationForProject.Project.KeyStage2PerformanceAdditionalInformation = additionalInformation;
+                FoundInformationForProject.Project.KeyStage4PerformanceAdditionalInformation = additionalInformation;
                 GetInformationForProject.Setup(s => s.Execute(ProjectUrn0001))
                     .ReturnsAsync(FoundInformationForProject);
 
@@ -67,7 +83,7 @@ namespace Frontend.Tests.PagesTests.TaskList
                 Assert.Equal(additionalInformation, _subject.AdditionalInformation.AdditionalInformation);
                 Assert.Equal(ProjectUrn0001, _subject.AdditionalInformation.Urn);
             }
-            
+
             [Fact]
             public async void GivenReturnToPreview_UpdatesTheViewModel()
             {
@@ -80,7 +96,7 @@ namespace Frontend.Tests.PagesTests.TaskList
             public async void GivenGetByUrnReturnsError_DisplayErrorPage()
             {
                 var pageModel =
-                    RazorPageTestHelpers.GetPageModelWithViewData<KeyStage2Performance>(
+                    RazorPageTestHelpers.GetPageModelWithViewData<KeyStage4Performance>(
                         GetInformationForProject.Object, ProjectRepository.Object);
 
                 var response = await pageModel.OnGetAsync(ProjectErrorUrn);
@@ -91,7 +107,7 @@ namespace Frontend.Tests.PagesTests.TaskList
             }
         }
 
-        public class OnPostAsyncTests : KeyStage2PerformanceTests
+        public class OnPostAsyncTests : KeyStage4PerformanceTests
         {
             private readonly Project _foundProject;
 
@@ -130,7 +146,7 @@ namespace Frontend.Tests.PagesTests.TaskList
             public async void GivenGetByUrnReturnsError_DisplayErrorPage()
             {
                 var pageModel =
-                    RazorPageTestHelpers.GetPageModelWithViewData<KeyStage2Performance>(
+                    RazorPageTestHelpers.GetPageModelWithViewData<KeyStage4Performance>(
                         GetInformationForProject.Object, ProjectRepository.Object);
 
                 var response = await pageModel.OnPostAsync(ProjectErrorUrn, string.Empty, false);
@@ -148,9 +164,9 @@ namespace Frontend.Tests.PagesTests.TaskList
                 var response = await _subject.OnPostAsync(ProjectUrn0001, additionalInformation, false);
 
                 var redirectToPageResponse = Assert.IsType<RedirectToPageResult>(response);
-                Assert.Equal("KeyStage2Performance", redirectToPageResponse.PageName);
+                Assert.Equal("KeyStage4Performance", redirectToPageResponse.PageName);
                 Assert.Equal("OnGetAsync", redirectToPageResponse.PageHandler);
-                Assert.Equal(additionalInformation, _foundProject.KeyStage2PerformanceAdditionalInformation);
+                Assert.Equal(additionalInformation, _foundProject.KeyStage4PerformanceAdditionalInformation);
             }
 
             [Fact]
@@ -160,10 +176,10 @@ namespace Frontend.Tests.PagesTests.TaskList
 
                 await _subject.OnPostAsync(ProjectUrn0001, additionalInfo, false);
                 ProjectRepository.Verify(r => r.Update(It.Is<Project>(
-                    project => project.KeyStage2PerformanceAdditionalInformation == additionalInfo
+                    project => project.KeyStage4PerformanceAdditionalInformation == additionalInfo
                 )));
             }
-            
+
             [Fact]
             public async void GivenReturnToPreview_RedirectsToThePreviewPage()
             {
