--- Frontend/Models/DownloadViewModel.cs
@@ -1,9 +1,0 @@
-﻿using Frontend.Models.Forms;
-
-namespace Frontend.Models
-{
-    public class DownloadViewModel : CommonViewModel
-    {
-        
-    }
-}
