+++ Frontend/Pages/TaskList/HtbDocument/Download.cshtml
@@ -1,4 +1,5 @@
-@model DownloadViewModel
+﻿@page "/project/{urn}/advisory-board/{handler?}"
+@model Frontend.Pages.TaskList.HtbDocument.Download
 
 @{
     ViewBag.Title = "Download project template";
@@ -22,14 +23,14 @@
         <section class="dfe-c-attachment">
             <div class="dfe-c-attachment__thumbnail">
                 <a class="govuk-link" target="_self" tabindex="-1" aria-hidden="true" href="#">
-                    <svg class="dfe-c-attachment__thumbnail-image" version="1.1" viewBox="0 0 99 140" width="99" height="140" aria-hidden="true">
+                    <svg class="dfe-c-attachment__thumbnail-image" viewBox="0 0 99 140" width="99" height="140" aria-hidden="true">
                         <path d="M12 12h75v27H12zM12 59h9v9h-9zM12 77h9v9h-9zM12 95h9v9h-9zM12 113h9v9h-9zM30 59h57v9H30zM30 77h39v9H30zM30 95h57v9H30zM30 113h48v9H30z" stroke-width="0"></path>
                     </svg>
                 </a>
             </div>
             <div class="dfe-c-attachment__details">
                 <h2 class="dfe-c-attachment__title">
-                    <a class="govuk-link govuk-link--no-visited-state" target="_blank" rel="noopener" download asp-action="GenerateDocument" asp-route-id="@Model.Urn" data-test="download-htb">
+                    <a class="govuk-link govuk-link--no-visited-state" target="_blank" rel="noopener" download asp-page-handler="GenerateDocument" asp-route-id="@Model.Urn" data-test="download-htb">
                         @Model.OutgoingAcademyName - Project template
                     </a>
                 </h2>
