+++ Frontend/Views/Project/Index.cshtml
@@ -126,7 +126,7 @@
             <form asp-page="@Links.HeadteacherBoard.Preview.PageName" asp-route-id="@Model.Project.Urn" method="get">
                 <button class="govuk-button" data-module="govuk-button" data-test="preview-htb">Preview project template</button>
             </form>
-            <form asp-controller="HeadteacherBoard" asp-action="Download" asp-route-id="@Model.Project.Urn" method="get">
+            <form asp-page="/Tasklist/HtbDocument/Download" asp-route-id="@Model.Project.Urn" method="get">
                 <button class="govuk-button govuk-button--secondary" data-module="govuk-button" data-test="generate-htb">Generate project template</button>
             </form>
         </div>
